+++ test/ex_check/project_cases/detect_test.exs
@@ -16,8 +16,8 @@ defmodule ExCheck.ProjectCases.DetectTest do
 
     assert {output, 0} = System.cmd("mix", ~w[check], cd: project_dir)
 
-    assert String.contains?(output, "credo skipped due to missing package credo")
-    assert String.contains?(output, "bad_dir skipped due to missing directory bad_directory")
-    refute String.contains?(output, "dialyzer skipped")
+    assert output =~ "credo skipped due to missing package credo"
+    assert output =~ "bad_dir skipped due to missing directory bad_directory"
+    refute output =~ "dialyzer skipped"
   end
 end