+++ test/ex_check/project_cases/external_tools_test.exs
@@ -6,12 +6,12 @@ defmodule ExCheck.ProjectCases.ExternalToolsTest do
 
     assert {output, 0} = System.cmd("mix", ~w[check], cd: project_dir, env: %{"MIX_ENV" => "dev"})
 
-    assert String.contains?(output, "compiler success")
-    assert String.contains?(output, "formatter success")
-    assert String.contains?(output, "ex_unit success")
-    assert String.contains?(output, "credo success")
-    assert String.contains?(output, "sobelow success")
-    assert String.contains?(output, "dialyzer skipped due to missing package dialyxir")
-    assert String.contains?(output, "ex_doc success")
+    assert output =~ "compiler success"
+    assert output =~ "formatter success"
+    assert output =~ "ex_unit success"
+    assert output =~ "credo success"
+    assert output =~ "sobelow success"
+    assert output =~ "dialyzer skipped due to missing package dialyxir"
+    assert output =~ "ex_doc success"
   end
 end