+++ test/ex_check/project_cases/gen_config_test.exs
@@ -1,7 +1,15 @@
-defmodule ExCheck.ProjectCases.BaseTest do
+defmodule ExCheck.ProjectCases.GenConfigTest do
   use ExCheck.ProjectCase, async: true
 
-  test "base", %{project_dir: project_dir} do
+  test "gen config", %{project_dir: project_dir} do
+    assert {output, 0} = System.cmd("mix", ~w[check.gen.config], cd: project_dir)
+
+    assert output =~ "creating .check.exs"
+
+    assert {output, 0} = System.cmd("mix", ~w[check.gen.config], cd: project_dir)
+
+    assert output =~ ".check.exs already exists, skipped"
+
     assert {output, 0} = System.cmd("mix", ~w[check], cd: project_dir)
 
     assert output =~ "compiler success"