+++ test/ex_check/umbrella_project_cases/base_test.exs
@@ -4,18 +4,18 @@ defmodule ExCheck.UmbrellaProjectCases.BaseTest do
   test "base", %{project_dirs: [project_root_dir | _]} do
     assert {output, 0} = System.cmd("mix", ~w[check], cd: project_root_dir)
 
-    assert String.contains?(output, "compiler success")
-    assert String.contains?(output, "formatter success")
-    refute String.contains?(output, "ex_unit success")
-    assert String.contains?(output, "ex_unit in child_a success")
-    assert String.contains?(output, "ex_unit in child_b success")
-    assert String.contains?(output, "credo skipped due to missing package credo")
-    refute String.contains?(output, "sobelow skipped due to missing package sobelow")
-    assert String.contains?(output, "sobelow in child_a skipped due to missing package sobelow")
-    assert String.contains?(output, "sobelow in child_b skipped due to missing package sobelow")
-    assert String.contains?(output, "dialyzer skipped due to missing package dialyxir")
-    assert String.contains?(output, "ex_doc skipped due to missing package ex_doc")
+    assert output =~ "compiler success"
+    assert output =~ "formatter success"
+    refute output =~ "ex_unit success"
+    assert output =~ "ex_unit in child_a success"
+    assert output =~ "ex_unit in child_b success"
+    assert output =~ "credo skipped due to missing package credo"
+    refute output =~ "sobelow skipped due to missing package sobelow"
+    assert output =~ "sobelow in child_a skipped due to missing package sobelow"
+    assert output =~ "sobelow in child_b skipped due to missing package sobelow"
+    assert output =~ "dialyzer skipped due to missing package dialyxir"
+    assert output =~ "ex_doc skipped due to missing package ex_doc"
 
-    assert String.contains?(output, "Randomized with seed")
+    assert output =~ "Randomized with seed"
   end
 end