+++ CHANGELOG.md
@@ -1,7 +1,9 @@
 # Changelog
 
-## v0.11.1-dev
+## v0.12.0-dev
 
+- **Added** `:deps` tool option with support for depending on specific exit status
+- **Removed** `:run_after` tool option (please use `:deps` tool option instead)
 - **Fixed** merging tool umbrella opts with those set in ancestor config
 - **Updated** default dialyxir config to no longer include `--halt-exit-status` deprecated in
   [1.0.0-rc.7](https://github.com/jeremyjh/dialyxir/blob/master/CHANGELOG.md#100-rc7---2019-09-21)